Rename core publishing operations
