+++ packages/core/src/publish/PublishCoreAttestation.ts
