+++ packages/core/src/publish/PublishCoreRemove.ts
@@ -1,6 +1,6 @@
 import { Connection } from "@solana/web3.js";
-import type { Publisher, SolanaMobileDappPublisherPortal } from "../types.js";
-import { createAttestationPayload } from "./attestation.js";
+import type { Publisher } from "../types.js";
+import { createAttestationPayload } from "./PublishCoreAttestation.js";
 import {
   CONTACT_OBJECT_ID,
   CONTACT_PROPERTY_COMPANY,
@@ -13,21 +13,17 @@ import {
   TICKET_PROPERTY_CRITICAL_UPDATE,
   TICKET_PROPERTY_DAPP_COLLECTION_ACCOUNT_ADDRESS,
   TICKET_PROPERTY_DAPP_RELEASE_ACCOUNT_ADDRESS,
-  TICKET_PROPERTY_POLICY_COMPLIANT,
   TICKET_PROPERTY_REQUEST_UNIQUE_ID,
-  TICKET_PROPERTY_TESTING_INSTRUCTIONS,
-  URL_FORM_UPDATE
+  URL_FORM_REMOVE
 } from "./dapp_publisher_portal.js";
 import { PublishSolanaNetworkInput, SignWithPublisherKeypair } from "./types.js";
 
-const createUpdateRequest = async (
+const createRemoveRequest = async (
   connection: Connection,
   sign: SignWithPublisherKeypair,
   appMintAddress: string,
   releaseMintAddress: string,
   publisherDetails: Publisher,
-  solanaMobileDappPublisherPortalDetails: SolanaMobileDappPublisherPortal,
-  compliesWithSolanaDappStorePolicies: boolean,
   requestorIsAuthorized: boolean,
   criticalUpdate: boolean
 ) => {
@@ -75,11 +71,6 @@ const createUpdateRequest = async (
         name: TICKET_PROPERTY_AUTHORIZED_REQUEST,
         value: requestorIsAuthorized
       },
-      {
-        objectTypeId: TICKET_OBJECT_ID,
-        name: TICKET_PROPERTY_POLICY_COMPLIANT,
-        value: compliesWithSolanaDappStorePolicies
-      }
     ]
   };
 
@@ -93,52 +84,36 @@ const createUpdateRequest = async (
     );
   }
 
-  if (solanaMobileDappPublisherPortalDetails.testing_instructions !== undefined) {
-    request.fields.push(
-      {
-        objectTypeId: TICKET_OBJECT_ID,
-        name: TICKET_PROPERTY_TESTING_INSTRUCTIONS,
-        value: solanaMobileDappPublisherPortalDetails.testing_instructions
-      }
-    );
-  }
-
   return request;
 };
 
-export type PublishUpdateInput = {
+export type PublishRemoveInput = {
   appMintAddress: string;
   releaseMintAddress: string;
   publisherDetails: Publisher;
-  solanaMobileDappPublisherPortalDetails: SolanaMobileDappPublisherPortal;
-  compliesWithSolanaDappStorePolicies: boolean;
   requestorIsAuthorized: boolean;
   criticalUpdate: boolean;
 };
 
-export const publishUpdate = async (
+export const publishRemove = async (
   publishSolanaNetworkInput: PublishSolanaNetworkInput,
   {
     appMintAddress,
     releaseMintAddress,
     publisherDetails,
-    solanaMobileDappPublisherPortalDetails,
-    compliesWithSolanaDappStorePolicies,
     requestorIsAuthorized,
     criticalUpdate,
-  } : PublishUpdateInput,
+  } : PublishRemoveInput,
   dryRun: boolean,
 ) => {
-  const updateRequest = await createUpdateRequest(
+  const removeRequest = await createRemoveRequest(
     publishSolanaNetworkInput.connection,
     publishSolanaNetworkInput.sign,
     appMintAddress,
     releaseMintAddress,
     publisherDetails,
-    solanaMobileDappPublisherPortalDetails,
-    compliesWithSolanaDappStorePolicies,
     requestorIsAuthorized,
     criticalUpdate);
 
-  submitRequestToSolanaDappPublisherPortal(updateRequest, URL_FORM_UPDATE, dryRun);
+  return submitRequestToSolanaDappPublisherPortal(removeRequest, URL_FORM_REMOVE, dryRun);
 };
