+++ packages/core/src/publish/index.ts
@@ -1,5 +1,5 @@
-export * from "./remove.js";
-export * from "./submit.js";
-export * from "./support.js";
+export * from "./PublishCoreRemove.js";
+export * from "./PublishCoreSubmit.js";
+export * from "./PublishCoreSupport.js";
 export * from "./types.js";
-export * from "./update.js";
+export * from "./PublishCoreUpdate.js";
